rename(file): contracts shortcuts
