+++ components/common/contracts-shortcuts.ts
